+++ ProcessManager.h
@@ -9,7 +9,7 @@ class ProcessManager
 {
 public:
     // Singleton stuff
-    static void initialize(int batch_process_freq, int min_ins, int max_ins);
+    static void initialize(long long batch_process_freq, long long min_ins, long long max_ins);
     static ProcessManager* getInstance();
 
     void createProcess(String name);
@@ -18,23 +18,23 @@ public:
     void displayAllProcesses();
     Process* findProcess(const String& name);
 
-    int getBatchProcessFreq() const;
-    int getMinInstructions() const;
-    int getMaxInstructions() const;
+    long long getBatchProcessFreq() const;
+    long long getMinInstructions() const;
+    long long getMaxInstructions() const;
     bool getIsGeneratingProcesses() const;
 
     void setIsGeneratingProcesses(bool val);
 
 private:
-    ProcessManager(int batch_process_freq, int min_ins, int max_ins);
+    ProcessManager(long long batch_process_freq, long long min_ins, long long max_ins);
     ProcessManager(const ProcessManager&);
     ProcessManager& operator = (const ProcessManager&);
     ~ProcessManager() = default;
     static ProcessManager* sharedInstance;
     
-    int batch_process_freq;
-    int min_ins;
-    int max_ins;
+    long long batch_process_freq;
+    long long min_ins;
+    long long max_ins;
     bool isGeneratingProcesses;
 
     std::vector<Process*> activeProcesses;
